Improved `Config` type definitions/documentation, re-generated baseline
Note: we did **not** remove redundant checks, since the primary purpose of the `Config`
object is to act as a small anti-corruption layer between user-provider configuration
and our own internal logic.

This may be changed in a new major version, but for now, reporting these redundant conditions
is a safer choice.